feat: auto-register slash commands on ready
Extract registration logic into register-commands.ts and call it in the
ClientReady handler so commands are always up to date on startup.
deploy-commands.ts is kept for manual runs.
